# [D] optimize matmul