--- a/src/d/matmul.d
+++ b/src/d/matmul.d
@@ -9,16 +9,15 @@ double[][] matgen(in int n) {
 	return a;
 }
 
-double[][] matmul(in double[][] a, in double[][] b) {
-	auto m = a.length, n = a[0].length, p = b[0].length;
-	auto c = new double[][](m, p);
-	for (int i = 0; i < m; ++i)
-		for (int j = 0; j < p; ++j)
+double[][] matmul(int n, in double[][] a, in double[][] b) {
+	auto c = new double[][](n, n);
+	for (int i = 0; i < n; ++i)
+		for (int j = 0; j < n; ++j)
 			c[i][j] = 0.0;
-	for (int i = 0; i < m; ++i) {
+	for (int i = 0; i < n; ++i) {
 		for (int k = 0; k < n; ++k) {
 			auto aik = a[i][k]; // faster this way
-			for (int j = 0; j < p; ++j)
+			for (int j = 0; j < n; ++j)
 				c[i][j] += aik * b[k][j];
 		}
 	}
@@ -29,6 +28,6 @@ void main(in string[] args) {
 	int n = 1500;
 	auto a = matgen(n);
 	auto b = matgen(n);
-	auto c = matmul(a, b);
+	auto c = matmul(n, a, b);
 	writeln(c[n / 2][n / 2]);
 }