Tested on Intel i5 10400 CPU, linux kernel: 6.6.8, ldc2: 1.35.0 (DMD v2.105.2, LLVM 16.0.6)

```
$ hyperfine -i --warmup 1 "./matmul_before" "./matmul_after"
Benchmark 1: ./matmul_before
  Time (mean ± σ):      3.578 s ±  0.017 s    [User: 3.564 s, System: 0.008 s]
  Range (min … max):    3.558 s …  3.614 s    10 runs
 
Benchmark 2: ./matmul_after
  Time (mean ± σ):      2.751 s ±  0.027 s    [User: 2.737 s, System: 0.008 s]
  Range (min … max):    2.712 s …  2.787 s    10 runs
 
Summary
  ./matmul_after ran
    1.30 ± 0.01 times faster than ./matmul_before

```

2nd commit even faster

```
$ hyperfine -i --warmup 1 "./matmul_after" "./matmul_after2"
Benchmark 1: ./matmul_after
  Time (mean ± σ):      2.742 s ±  0.024 s    [User: 2.734 s, System: 0.005 s]
  Range (min … max):    2.712 s …  2.786 s    10 runs
 
Benchmark 2: ./matmul_after2
  Time (mean ± σ):      1.502 s ±  0.024 s    [User: 1.493 s, System: 0.006 s]
  Range (min … max):    1.457 s …  1.541 s    10 runs
 
Summary
  ./matmul_after2 ran
    1.83 ± 0.03 times faster than ./matmul_after

```